move the api into stats
