+++ pkg/api/dashboard.go
@@ -8,12 +8,10 @@ import (
 	"net/http"
 	"os"
 	"path/filepath"
-	"slices"
 	"strconv"
 	"strings"
 
 	"github.com/grafana/authlib/claims"
-	"github.com/prometheus/prometheus/promql/parser"
 
 	"github.com/grafana/grafana/pkg/api/apierrors"
 	"github.com/grafana/grafana/pkg/api/dtos"
@@ -235,99 +233,6 @@ func (hs *HTTPServer) GetDashboard(c *contextmodel.ReqContext) response.Response
 
 	c.TimeRequest(metrics.MApiDashboardGet)
 	return response.JSON(http.StatusOK, dto)
-}
-
-type DashboardMetricStats struct {
-	MetricStats map[string]int `json:"metric-stats"`
-}
-
-func (hs *HTTPServer) GetMetricStats(c *contextmodel.ReqContext) response.Response {
-	ctx, span := tracer.Start(c.Req.Context(), "api.GetMetricStats")
-	defer span.End()
-	c.Req = c.Req.WithContext(ctx)
-
-	dashboardQuery := dashboards.GetExtractedExpressionsQuery{OrgID: c.SignedInUser.GetOrgID(), DSType: "prometheus"}
-	dashboardQueryResult, err := hs.DashboardService.GetDashboardMetricStats(c.Req.Context(), &dashboardQuery)
-	if err != nil {
-		c.JsonApiErr(500, "failed to get exprs from dashboards", err)
-		return response.Err(fmt.Errorf("failed to get exprs from dashboards: %v", err))
-	}
-
-	alertRuleQuery := dashboards.GetExtractedExpressionsQuery{OrgID: c.SignedInUser.GetOrgID(), DSType: "prometheus"}
-	alertRuleQueryResult, err := hs.DashboardService.GetAlertRuleMetricStats(c.Req.Context(), &alertRuleQuery)
-	if err != nil {
-		c.JsonApiErr(500, "failed to get exprs from dashboards", err)
-		return response.Err(fmt.Errorf("failed to get exprs from dashboards: %v", err))
-	}
-
-	metricCount := map[string]map[string]int{
-		"alert_rules": make(map[string]int),
-		"dashboards":  make(map[string]int),
-	}
-
-	countMetricNames(alertRuleQueryResult, metricCount["alert_rules"])
-	countMetricNames(dashboardQueryResult, metricCount["dashboards"])
-
-	return response.JSON(http.StatusOK, metricCount)
-}
-
-func countMetricNames(queryResult []*dashboards.ExtractedExpression, metricCount map[string]int) {
-	uniqueMetricName := make(map[string]int64)
-
-	// Iterate over each expression and extract metrics
-	for _, expression := range queryResult {
-		metricNames, err := extractMetricName(expression.Expr)
-		if err != nil {
-			fmt.Println("couldn't extract the metric name.")
-			continue
-		}
-		// Count each metric occurrence only once for each dashboard
-		for _, metric := range metricNames {
-			if metric != "" {
-				if _, processed := uniqueMetricName[metric]; !processed || uniqueMetricName[metric] != expression.ID {
-					uniqueMetricName[metric] = expression.ID
-					metricCount[metric]++
-				}
-			}
-		}
-	}
-}
-
-// MetricVisitor is a custom Visitor to collect metric names from a PromQL AST
-type MetricVisitor struct {
-	metrics []string
-}
-
-// Visit is the method that handles the AST nodes and extracts metric names
-func (v *MetricVisitor) Visit(node parser.Node, path []parser.Node) (parser.Visitor, error) {
-	switch n := node.(type) {
-	case *parser.VectorSelector:
-		// For a VectorSelector, it's the metric name
-		if !slices.Contains(v.metrics, n.Name) {
-			v.metrics = append(v.metrics, n.Name)
-		}
-	}
-	// Continue walking the tree
-	return v, nil
-}
-
-// Function to extract metric names from a PromQL expression
-func extractMetricName(expr string) ([]string, error) {
-	// Parse the PromQL expression into an AST
-	node, err := parser.ParseExpr(expr)
-	if err != nil {
-		return nil, err
-	}
-
-	// Create a visitor and walk the AST
-	visitor := &MetricVisitor{}
-	err = parser.Walk(visitor, node, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	// Return the collected metric names
-	return visitor.metrics, nil
 }
 
 func (hs *HTTPServer) getAnnotationPermissionsByScope(c *contextmodel.ReqContext, actions *dashboardsV0.AnnotationActions, scope string) {
