+++ pkg/services/dashboards/database/database.go
@@ -950,79 +950,6 @@ func (d *dashboardStore) GetDashboardTags(ctx context.Context, query *dashboards
 	return queryResult, nil
 }
 
-func (d *dashboardStore) GetExpressionsFromDashboardPanels(ctx context.Context, query *dashboards.GetExtractedExpressionsQuery) ([]*dashboards.ExtractedExpression, error) {
-	ctx, span := tracer.Start(ctx, "dashboards.database.GetExpressionsFromDashboardPanels")
-	defer span.End()
-
-	queryResult := make([]*dashboards.ExtractedExpression, 0)
-	err := d.store.WithDbSession(ctx, func(dbSession *db.Session) error {
-		sql := fmt.Sprintf(`WITH RECURSIVE
-				    panel_targets AS (
-				        SELECT
-				            dashboard.id,
-				            json_extract(target.value, '$.expr') AS expr
-				        FROM
-				            dashboard,
-				            json_each(dashboard.data, '$.panels') AS panel,
-				            json_each(panel.value, '$.targets') AS target
-				        WHERE
-				            json_extract(target.value, '$.datasource.type') = '%s'
-				    )
-				SELECT
-				    id,
-				    expr
-				FROM
-				    panel_targets
-				WHERE
-				    expr IS NOT NULL AND trim(expr) != ''`, query.DSType)
-
-		sess := dbSession.SQL(sql, query.OrgID)
-		err := sess.Find(&queryResult)
-		return err
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return queryResult, nil
-}
-
-func (d *dashboardStore) GetExpressionsFromAlertRules(ctx context.Context, query *dashboards.GetExtractedExpressionsQuery) ([]*dashboards.ExtractedExpression, error) {
-	ctx, span := tracer.Start(ctx, "dashboards.database.GetExpressionsFromAlertRules")
-	defer span.End()
-
-	queryResult := make([]*dashboards.ExtractedExpression, 0)
-	err := d.store.WithDbSession(ctx, func(dbSession *db.Session) error {
-		sql := fmt.Sprintf(`WITH RECURSIVE
-				    alert_rule_targets AS (
-				        SELECT
-				            alert_rule.id,
-				            json_extract(rule.value, '$.model.expr') AS expr
-				        FROM
-				            alert_rule,
-				            json_each(alert_rule.data, '$') AS rule  -- Iterate over the array of rules directly
-				        WHERE
-				            json_extract(rule.value, '$.model.datasource.type') = '%s'
-				    )
-				SELECT
-				    id,
-				    expr
-				FROM
-				    alert_rule_targets
-				WHERE
-				    expr IS NOT NULL AND trim(expr) != ''`, query.DSType)
-
-		sess := dbSession.SQL(sql, query.OrgID)
-		err := sess.Find(&queryResult)
-		return err
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return queryResult, nil
-}
-
 // CountDashboardsInFolder returns a count of all dashboards associated with the
 // given parent folder ID.
 func (d *dashboardStore) CountDashboardsInFolders(
